In this update I added 2 more hotkeys; numpad 5 performs reset search and numpad 6 performs search.
I also updated the set location, yes, no and pick folder to use the latest custom toast instead of using the default toast, and I also updated the messages for the yes and no button clicks to make them look better.

And for the final update, I updated the GUI for the GetSearchNameFromUser form since I personally didn't think it looked nice and didn't look consistent with the dark colours of the main form.